% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_supply_and_demand}
\alias{create_supply_and_demand}
\title{Create System of Supply and Demand}
\usage{
create_supply_and_demand(qd = 1000, qs = 250, a = 10, b = 5)
}
\arguments{
\item{qd}{exogenous demand.}

\item{qs}{exogenous supply. Default: 250}

\item{a}{Demand's sensitivity to the price. Default: 10}

\item{b}{Supply's sensitivity to the price. Default: 5}
}
\value{
A tibble that represents a system of linear equations for supply and demand.
}
\description{
Creates a tibble to represent a system of linear equations for supply and demand.
}
\details{
Use it to create a tibble and then use it with plot_system_and_demand.
}
\examples{
\dontrun{
if(interactive()){
 create_supply_and_demand() \%>\%
 plot_supply_and_demand()
 }
}
}
\seealso{
\code{\link[Hmisc]{bezier}}
 \code{\link[tibble]{as_data_frame}}
 \code{\link[dplyr]{rename}},\code{\link[dplyr]{left_join}}
}
