% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{curve_intersect_two}
\alias{curve_intersect_two}
\title{Find Intersection}
\usage{
curve_intersect_two(df)
}
\arguments{
\item{df}{A matrix with the curves approximated numerically.}
}
\value{
A tibble with the point of intersection between the curves.
}
\description{
Given a dataframe with two curves (first the x) and then the output of functions, it approximates the intersection between the functions.
}
\details{
This function was taken directly from https://www.andrewheiss.com/blog/2017/09/15/create-supply-and-demand-economics-curves-with-ggplot2/
}
