% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_all_intersections}
\alias{find_all_intersections}
\title{Find Curve Intersections}
\usage{
find_all_intersections(supply_and_demand_system)
}
\arguments{
\item{supply_and_demand_system}{A tibble that represents a system of linear equations for supply and demand.}
}
\value{
A tibble with the respective curves and their intersection.
}
\description{
Given a tibble representing a system of supply and demand, find
  all the intersections between the supply curves and demand curves.
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{bind_rows}}
}
