% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shift_curve}
\alias{shift_curve}
\title{Curve Shifter}
\usage{
shift_curve(supply_and_demand_system, curve, outwards = TRUE)
}
\arguments{
\item{supply_and_demand_system}{A tibble created with \code{create_supply_and_demand()}}

\item{curve}{The curve that you want to shift from the tibble.}

\item{outwards}{Whether you want to shift the curve outwards or not. Default: TRUE}
}
\value{
A tibble with an added curve that results from the shift.
}
\description{
Shift one of the curves in the tibble from \code{create_supply_and_demand()}.
}
\details{
Not exposed to the user. Just for internal use in the programming.
}
\seealso{
\code{\link[stringr]{str_detect}}
 \code{\link[dplyr]{mutate}}
 \code{\link[rlang]{quo_name}}
}
