% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shift_demand}
\alias{shift_demand}
\title{Shift Demand}
\usage{
shift_demand(supply_and_demand_system, outwards = TRUE, curve = demand)
}
\arguments{
\item{supply_and_demand_system}{A tibble created from \code{create_supply_and_demand()}}

\item{outwards}{Whether you want to shift the demand outwards or not. Default: TRUE}

\item{curve}{Which of the existing demands you want to shift: i.e., shift a demand already shifted \code{demand1}, Default: demand}
}
\value{
A tibble with an extra demand curve resulting from the shifting.
}
\description{
Takes a tibble that represents a system of supply and demand and
  creates another curve by shifting one of the already existing demands.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 create_supply_and_demand() \%>\%
 shift_demand() \%>\%
 shift_demand(demand = demand1) \%>\%
 plot_supply_and_demand()
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}
}
